backend/app/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pyproj
shapely
supabase
python-jose[cryptography]
passlib[bcrypt]
email-validator

# Optional: Local routing fallback dependencies (install on-demand; on Windows prefer conda)
osmnx>=1.3
networkx
